requirements.txt
BeautifulSoup==3.2.1
Flask==0.9
Jinja2==2.6
PyPitches==0.1.0
PyYAML==3.10
SQLAlchemy==0.7.10
Werkzeug==0.8.3
argparse==1.2.1
ipython==0.13.1
psycopg2==2.4.6
python-dateutil==2.1
six==1.2.0
wsgiref==0.1.2
matplotlib==1.2.0
numpy==1.6.2
nose==1.2.1
